+++ src/controller/clientesController.js
@@ -3,7 +3,8 @@ const Cliente = require('../Model/clientesModel.js');
 
 exports.listarClientes = async (req, res) => {
     const clientes = await Cliente.findAll();
-    res.render(clientes); // trocar por Render quando for usar EJS e JSON para teste em HTTP ou ISOMNIA
+    res.render('consultaCliente'); // trocar por Render quando for usar EJS e JSON para teste em HTTP ou ISOMNIA
+  //  res.status(200).json(clientes);
 };
 
 exports.adicionarCliente = async (req, res) => {
